+++ src/optilab/functions/surrogate/xgboost_surrogate_objective_function.py
@@ -63,8 +63,8 @@ class XGBoostSurrogateObjectiveFunction(SurrogateObjectiveFunction):
         super().train(train_set)
 
         x_train, y_train = self.train_set.pairs()
-        x_train = np.array(x_train, dtype=np.float32)
-        y_train = np.array(y_train, dtype=np.float32)
+        x_train = np.array(x_train, dtype=np.float64)
+        y_train = np.array(y_train, dtype=np.float64)
 
         self.model = xgb.XGBRegressor(
             n_estimators=self.n_estimators,
@@ -86,7 +86,7 @@ class XGBoostSurrogateObjectiveFunction(SurrogateObjectiveFunction):
         """
         super().__call__(point)
 
-        x_query = np.array([point.x], dtype=np.float32)
+        x_query = np.array([point.x], dtype=np.float64)
         y_pred = self.model.predict(x_query)[0]
 
         return Point(
